erfactor suggestion hooks, separate single and multiple mode
