+++ vuu-ui/packages/parsed-input/src/parsed-input/suggestions/index.js
@@ -1,0 +1,2 @@
+export * from './SuggestionList';
+export * from './use-suggestions';
